# backend/clients/ts-sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Patron API",
    "description": "An open source Patreon alternative with lower fees designed for creators who publish ongoing sequential content like books, podcasts, and comics.",
    "contact": {
      "name": "Patron Team",
      "email": "support@patron.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.patron.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/cdn/files/{file_id}": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "Serve file content with user authentication",
        "description": "This endpoint is designed to be used to get file content with proper authentication.\nIt verifies user access to the file and returns the file content with proper cache headers.\nThe file content is streamed directly from S3 to minimize memory usage for large files.\n\n# Errors\nReturns an error if file not found, access denied, or S3 operations fail.",
        "operationId": "serve_file_cdn",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "description": "UUID of the file to stream via CDN",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Streaming file content with CDN-optimized headers",
            "content": {
              "application/octet-stream": {
                "example": "Binary file content with appropriate Content-Type and Cache-Control headers"
              }
            }
          },
          "401": {
            "description": "Authentication required for CDN file access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "CDN access forbidden - user cannot access this file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "CDN file not found or access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "CDN streaming error from storage backend",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/files": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "List user's files with cursor-based pagination",
        "description": "# Errors\nReturns an error if database operations fail.",
        "operationId": "list_files",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "description": "UUID offset for cursor-based files pagination",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of files to return (default: 50, max: 100)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User files retrieved with pagination support",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserFilesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required to access files",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database error while retrieving user files",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/files/actions/upload": {
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Upload a file",
        "description": "# Errors\nReturns an error if file upload, database operations, or file system operations fail.",
        "operationId": "upload_file",
        "requestBody": {
          "description": "Multipart form data with file field",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadRequest"
              },
              "example": {
                "file": "Binary file content - upload any file type (PDF, image, document, etc.)"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "File uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid file upload request or malformed file data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required for file upload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Uploaded file exceeds maximum size limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error during file upload or storage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard JSON error response structure for API endpoints.",
        "required": [
          "error"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional error code for programmatic handling",
            "example": "AUTH_INVALID_CREDENTIALS"
          },
          "error": {
            "type": "string",
            "description": "Error message describing what went wrong",
            "example": "Invalid email or password"
          }
        },
        "example": {
          "code": "AUTH_INVALID_CREDENTIALS",
          "error": "Invalid email or password"
        }
      },
      "FileStatus": {
        "type": "string",
        "description": "File processing status for user uploaded files",
        "enum": [
          "uploaded",
          "processing",
          "processed",
          "failed",
          "deleted"
        ]
      },
      "FileUploadRequest": {
        "type": "object",
        "description": "File upload request schema for multipart form data",
        "required": [
          "file"
        ],
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "description": "The file to upload",
            "example": "Binary file content (PDF, image, document, etc.)"
          }
        }
      },
      "FileUploadResponse": {
        "type": "object",
        "description": "Response for file upload operations",
        "required": [
          "message",
          "file"
        ],
        "properties": {
          "file": {
            "$ref": "#/components/schemas/UserFileInfo",
            "description": "Uploaded file information"
          },
          "message": {
            "type": "string",
            "description": "Upload confirmation message"
          }
        },
        "example": {
          "file": {
            "createdAt": "2023-01-01T00:00:00Z",
            "fileHash": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
            "fileSize": 1048576,
            "filename": "document.pdf",
            "id": "d290f1ee-6c54-4b01-90e6-d701748f0851",
            "metadata": {
              "description": "Important document"
            },
            "mimeType": "application/pdf",
            "originalFilename": "My Important Document.pdf",
            "status": "uploaded",
            "updatedAt": "2023-01-01T00:00:00Z",
            "userId": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
          },
          "message": "File uploaded successfully"
        }
      },
...
      "UserFileInfo": {
        "type": "object",
        "description": "User file information for API responses",
        "required": [
          "id",
          "userId",
          "filename",
          "originalFilename",
          "fileSize",
          "mimeType",
          "fileHash",
          "status"
        ],
        "properties": {
          "createdAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "File upload timestamp",
            "example": "2023-01-01T00:00:00Z"
          },
          "fileHash": {
            "type": "string",
            "description": "SHA-256 hash of the file",
            "example": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64",
            "description": "Size of the file in bytes",
            "example": 1048576
          },
          "filename": {
            "type": "string",
            "description": "Current filename",
            "example": "document.pdf"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "File's unique identifier",
            "example": "b8c9d0e1-2345-6789-b234-678901234567"
          },
          "metadata": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Value",
                "description": "Additional file metadata"
              }
            ]
          },
          "mimeType": {
            "type": "string",
            "description": "MIME type of the file",
            "example": "application/pdf"
          },
          "originalFilename": {
            "type": "string",
            "description": "Original filename as uploaded by user",
            "example": "My Important Document.pdf"
          },
          "status": {
            "$ref": "#/components/schemas/FileStatus",
            "description": "Current processing status"
          },
          "updatedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "File last update timestamp",
            "example": "2023-01-01T00:00:00Z"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the user who owns this file",
            "example": "c9d0e1f2-3456-789a-c345-789012345678"
          }
        },
        "example": {
          "createdAt": "2023-01-01T00:00:00Z",
          "fileHash": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
          "fileSize": 1048576,
          "filename": "document.pdf",
          "id": "b8c9d0e1-2345-6789-b234-678901234567",
          "metadata": {
            "height": 1080,
            "width": 1920
          },
          "mimeType": "application/pdf",
          "originalFilename": "My Important Document.pdf",
          "status": "uploaded",
          "updatedAt": "2023-01-01T00:00:00Z",
          "userId": "c9d0e1f2-3456-789a-c345-789012345678"
        }
      },
      "UserFilesResponse": {
        "type": "array",
        "title": "UserFilesResponse",
        "items": {
          "$ref": "#/components/schemas/UserFileInfo"
        },
        "description": "Collection of user files with metadata",
        "example": [
          {
            "createdAt": "2023-01-01T00:00:00Z",
            "fileHash": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
            "fileSize": 1048576,
            "filename": "document.pdf",
            "id": "b8c9d0e1-2345-6789-b234-678901234567",
            "metadata": {
              "description": "Important document"
            },
            "mimeType": "application/pdf",
            "originalFilename": "My Important Document.pdf",
            "status": "uploaded",
            "updatedAt": "2023-01-01T00:00:00Z",
            "userId": "c9d0e1f2-3456-789a-c345-789012345678"
          }
        ]
      },
...
      "Value": {
        "type": "object",
        "title": "JSONValue",
        "description": "Flexible JSON value that can be an object, array, string, number, boolean, or null"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API key authentication using Bearer token in Authorization header"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid",
        "description": "Session-based authentication using secure HTTP-only cookies"
      }
    }
  },
...
}